升级版本到 2.0.7，修复 IDE 兼容性（until-build 263.*）
- plugin.xml: 添加 until-build="263.*" 支持 IDE 261+
- build.gradle: 版本 2.0.5→2.0.7, untilBuild 252.*→263.*, Kotlin 2.1.0→2.3.0
- CodeLocatorViewNodeDecoder.kt: 修复 Byte.toChar() 弃用错误
- DeleteImportManager.kt: 修复 OptimizeImportsProcessor 在 write action 内执行的问题
- 添加根目录 Gradle Wrapper 和构建配置
- 更新 .gitignore 排除 local.properties、.gradle/、build/

+++ CodeLocatorPlugin/src/main/java/com/bytedance/tools/codelocator/parser/CodeLocatorViewNodeDecoder.kt
@@ -59,7 +59,7 @@ class CodeLocatorViewNodeDecoder(private val mBuf: ByteBuffer) {
         constructor(
             seen: Byte,
             pos: Int
-        ) : super(String.format("Unexpected byte %c seen at position %d", seen.toChar(), pos))
+        ) : super(String.format("Unexpected byte %c seen at position %d", seen.toInt().toChar(), pos))
 
         constructor(msg: String) : super(msg)
     }
